requirements.txt
uvicorn==0.32.0
fastapi==0.115.2
python-dotenv==1.0.1
huggingface-hub==0.25.2
transformers==4.45.2
torch==2.4.1
pydantic==2.9.2
sentence-transformers==3.3.0